refactor: extract composables from App.vue
Extract six composables to reduce App.vue from 1601 to 1032 lines
(script section from ~660 to ~90 lines of wiring):

- useWeekNavigation: week navigation state, date formatting, prev/next/today
- useWeekPersistence: week data loading, saving, debounced save
- useCustomColumns: custom columns CRUD, persistence, data application
- useRecurringTasks: materialization and cleanup of recurring task instances
- useTaskEditing: edit dialog state, save/delete, recurrence handling, task CRUD
- usePolling: long-polling and notify_push integration

No behavioral changes. Build passes clean.
